+++ lib/src/components/form/fields/select.dart
@@ -62,6 +62,9 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
 
     /// {@macro ShadSelect.shrinkWrap}
     bool? shrinkWrap,
+
+    /// {@macro ShadSelect.controller}
+    ShadSelectController<T>? controller,
   }) : super(
           decorationBuilder: (context) =>
               (ShadTheme.of(context).selectTheme.decoration ??
@@ -100,6 +103,7 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
               groupId: groupId,
               itemCount: itemCount,
               shrinkWrap: shrinkWrap,
+              controller: controller,
             );
           },
         );
@@ -164,6 +168,9 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
 
     /// {@macro ShadSelect.shrinkWrap}
     bool? shrinkWrap,
+
+    /// {@macro ShadSelect.controller}
+    ShadSelectController<T>? controller,
   }) : super(
           decorationBuilder: (context) =>
               (ShadTheme.of(context).selectTheme.decoration ??
@@ -209,6 +216,7 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
               groupId: groupId,
               itemCount: itemCount,
               shrinkWrap: shrinkWrap,
+              controller: controller,
             );
           },
         );
@@ -270,6 +278,9 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
 
     /// {@macro ShadSelect.shrinkWrap}
     bool? shrinkWrap,
+
+    /// {@macro ShadSelect.controller}
+    ShadSelectController<T>? controller,
   })  : assert(
           variant == ShadSelectVariant.primary || onSearchChanged != null,
           'onSearchChanged must be provided when variant is search',
@@ -323,6 +334,7 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
               groupId: groupId,
               itemCount: itemCount,
               shrinkWrap: shrinkWrap,
+              controller: controller,
             );
           },
         );
@@ -382,6 +394,9 @@ class ShadSelectMultipleFormField<T> extends ShadFormBuilderField<List<T>> {
 
     /// {@macro ShadSelect.allowDeselection}
     bool allowDeselection = true,
+
+    /// {@macro ShadSelect.controller}
+    ShadSelectController<T>? controller,
   }) : super(
           decorationBuilder: (context) =>
               (ShadTheme.of(context).selectTheme.decoration ??
@@ -396,7 +411,7 @@ class ShadSelectMultipleFormField<T> extends ShadFormBuilderField<List<T>> {
               selectedOptionsBuilder: selectedOptionsBuilder,
               focusNode: state.focusNode,
               placeholder: placeholder,
-              initialValues: initialValue ?? const [],
+              initialValues: initialValue ?? [],
               enabled: state.enabled,
               onChanged: state.didChange,
               closeOnTapOutside: closeOnTapOutside,
@@ -417,6 +432,7 @@ class ShadSelectMultipleFormField<T> extends ShadFormBuilderField<List<T>> {
               footer: footer,
               closeOnSelect: closeOnSelect,
               allowDeselection: allowDeselection,
+              controller: controller,
             );
           },
         );
@@ -474,6 +490,9 @@ class ShadSelectMultipleFormField<T> extends ShadFormBuilderField<List<T>> {
 
     /// {@macro ShadSelect.allowDeselection}
     bool allowDeselection = true,
+
+    /// {@macro ShadSelect.controller}
+    ShadSelectController<T>? controller,
   }) : super(
           decorationBuilder: (context) =>
               (ShadTheme.of(context).selectTheme.decoration ??
@@ -488,7 +507,7 @@ class ShadSelectMultipleFormField<T> extends ShadFormBuilderField<List<T>> {
               selectedOptionsBuilder: selectedOptionsBuilder,
               focusNode: state.focusNode,
               placeholder: placeholder,
-              initialValues: initialValue ?? const [],
+              initialValues: initialValue ?? [],
               enabled: state.enabled,
               onChanged: state.didChange,
               closeOnTapOutside: closeOnTapOutside,
@@ -516,6 +535,7 @@ class ShadSelectMultipleFormField<T> extends ShadFormBuilderField<List<T>> {
               footer: footer,
               closeOnSelect: closeOnSelect,
               allowDeselection: allowDeselection,
+              controller: controller,
             );
           },
         );
@@ -568,6 +588,9 @@ class ShadSelectMultipleFormField<T> extends ShadFormBuilderField<List<T>> {
     Widget? footer,
     bool allowDeselection = true,
     bool closeOnSelect = true,
+
+    /// {@macro ShadSelect.controller}
+    ShadSelectController<T>? controller,
   })  : assert(
           variant == ShadSelectVariant.multiple ||
               variant == ShadSelectVariant.multipleWithSearch,
@@ -587,7 +610,7 @@ class ShadSelectMultipleFormField<T> extends ShadFormBuilderField<List<T>> {
               selectedOptionsBuilder: selectedOptionsBuilder,
               focusNode: state.focusNode,
               placeholder: placeholder,
-              initialValues: initialValue ?? const [],
+              initialValues: initialValue ?? [],
               enabled: state.enabled,
               onMultipleChanged: state.didChange,
               closeOnTapOutside: closeOnTapOutside,
@@ -614,6 +637,7 @@ class ShadSelectMultipleFormField<T> extends ShadFormBuilderField<List<T>> {
               footer: footer,
               allowDeselection: allowDeselection,
               closeOnSelect: closeOnSelect,
+              controller: controller,
             );
           },
         );